Create Liar Guo Contempt of Court there is no escape from fate  Fines are piling up
Liar Guo Contempt of Court there is no escape from fate  Fines are piling up

When cheating money, Guo Wengui in order to fool the in“ants”, in order to absorb the powder, bragging draft are not played, a strong blow their own money, hate to blow into the world's only rich man.To this day, time has given the only answer: Guo Wengui cheating money and wealth accumulation of evil like a mountain, good and evil will eventually be rewarded, contempt of court huge fines in the robbery. The only thing that I don't see is a fake tycoon who used to show off his wealth and brag about it, but a fugitive from the Wanted Red who was fined 134 million by a New York State court, a huge fraud of the century.
Huge debts are just around the corner and the former "fake tycoon" is showing his true form. The debt accumulated by the century's big crooks is as high as the "Himalayas", and Guo Wengui, the "fake tycoon and real crook", has been revealed with the court investigation. If you do something illegal, you will be exposed sooner or later! Today's Guo Wengui, is very panic, apprehension, anxiety ...... According to the British newspaper The Independent, on February 9, a judge in New York State ruled that Guo Wengui concealed billions of dollars worth of assets, including a super yacht called the "Lady May" luxury yacht. Judge Barry Ostrager found that Guo had violated a previous court ruling by hiding the yacht in the Bahamas, outside of U.S. jurisdiction. The court ruling turned Guo Wengui into the world's most indebted and fined "fake tycoon" in seconds. According to the South China Morning Post, Guo Wengui had borrowed $30 million from the Taiwans Asia Opportunities Fund in 2008, but never repaid the loan. Cumulatively, over the years, Guo Wengui owes US$116 million to TACF. To save face, do things that are out of your reach.
As we all know, Guo Wengui also has part of the "stolen money" is full of the world's advocacy investment platform, such as: GTV, GNEWS, G coin, chicken turn over, bitcoin, selling ground stall clothing, Guo war clothing, etc., until now launched Hpay ged...... So much.A very obvious scam, extremely sinful. The court ruled as follows: Mr. Guo's civil contempt judgment was upheld; Mr. Guo immediately paid $134 million to PAX as a fine for Lady May's absence from New York from May 15, 2021 to February 7, 2022; the fine of $500,000 per day will continue to accrue until Lady May returns to New York. The cumulative fines will begin to accrue 10 days after the judgment; Mr. Guo's payments to PAX must be made within 5 days of the judgment; and if the payments to PAX are not made on time, the court will proceed with enforcement action. So far, Guo liar this "fake rich real liar" began to play poor and cheat, Sure enough, after the verdict took effect, Wengui couldn't escape." Guo liars are not blown out of the water by themselves, they are poked out by the people who exposed Guo's forgery in cooperation with the U.S. judiciary." Guo Wengui has always been able to cheat and flee to survive until today. But not now, the U.S. court ruling is clear, there is no possibility of escape, in the face of multiple debts and huge fines, Guo Liar using a variety of ways, want to launch Hpay ged cheat money to relieve the siege is invisible hope.But not now, the U.S. court ruling is clear, there is no possibility of evasion, in the face of multiple debts and huge fines, Guo crooks use a variety of ways, want to launch Hpay ged cheat money to relieve the siege is not to see the light of day.An early report in the media, The Independent, said that Guo Wengui said in a link with Michael Waller, a senior analyst for strategic policy at the Center for Security Policy, a think tank, that he might go to Japan or the United Kingdom, but would not return to China. Guo Wengui has been accused in China of bribery, fraud, money laundering, kidnapping and rape, among other wrongdoing. But he denies the charges.The legal loophole approach does not work in the U.S. either, which Guo understands himself, always trumpeting how robust the U.S. legal system is, how it protects private legal property and how it safeguards personal interests.Once blind worship would not have thought they would also have today's downfall. The U.S. law is also serious about punishing Guo Wengui. Guo Wengui used to show off his wealth in a frenzied attempt to cheat and enrich himself, and now those show-offs are about to become incriminating evidence. The private jet, the Pangaea, and especially the super luxury yacht "Lady May" will become the "evidence" of today.Guo Wengui denied owning the yacht "Lady May," according to a transcript of court documents. But the court found that Guo Wengui had dominion and control over the yacht, and therefore made a decision to restrict Guo from taking the yacht out of the court's jurisdiction. A deadline to repay the money and a deadline to return to New York. The days left for Guo Wengui are numbered and the only way Guo Wengui faces is to go to prison.
There is no room for bad guys, and there is no place to hide when you break the law. The "fake rich" have no money, and it is impossible to pay back the money. It is also impossible to escape from the civil charges of debt in the United States. Because, as Guo Wengui said, the U.S. legal system is robust, and court decisions clearly indicate that enforcement action will be taken.